requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
redis
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
httpx
pytest
firebase-admin